refactor(sign.js) : 여러 층위의 모듈로 분리해보는 연습
