+++ signin.html
@@ -52,6 +52,6 @@
         </footer>
     </main>
 
-    <script type="text/javascript" src="/asset/js/sign.js"></script>
+    <script type="module" src="/asset/js/main.js"></script>
 </body>
 </html>
